requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch>=2.1.2+cu118
torchaudio>=2.1.2+cu118
torchvision>=0.16.2+cu118
hydra-core==1.3.2
numpy==1.26.4
epochalyst>=0.1.5
scikit-learn~=1.4.1
wandb>=0.16.4
coloredlogs==15.0.1
dask-expr==1.0.3
tqdm==4.66.2
annotated_types==0.6.0
timm==0.9.16
randomname==0.2.1
pyarrow==15.0.0
kaggle==1.6.6

-e ./
